Move EthBlocks::block_receipts default impl to impl for EthApi
